Add and order book info
This commit shouldn't be on this branch, but oh well
